requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --output-file requirements.txt --no-hashes
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   starlette
    #   watchfiles
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
    #   uvicorn
fastapi==0.121.1
    # via gs-ia
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.2 ; python_full_version >= '3.14'
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-core==2.28.1 ; python_full_version < '3.14'
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.43.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-generativeai==0.8.5
    # via gs-ia
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via uvicorn
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.4
    # via
    #   fastapi
    #   google-generativeai
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.2.5
    # via httplib2
python-decouple==3.8
    # via gs-ia
python-dotenv==1.2.1
    # via uvicorn
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via google-api-core
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.49.3
    # via fastapi
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via gs-ia
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
